[WII WAD UNPACKER] Working.  Need to add U8 unpacker and auto U8 unpack for WAD unpacker.

+++ vgmtutil/VGMToolbox/util/FileUtil.cs
@@ -201,7 +201,7 @@ namespace VGMToolbox.util
             }
         }
 
-        public static void TrimFileToLength(string path, int totalLength)
+        public static void TrimFileToLength(string path, long totalLength)
         {
             string fullPath = Path.GetFullPath(path);
